Add more documentation to the example app
