+++ examples/monitoring/main.go
@@ -8,30 +8,11 @@ import (
 	"syscall"
 	"time"
 
-	"github.com/sirupsen/logrus"
-
-	"github.com/capatazlib/go-capataz/c"
 	"github.com/capatazlib/go-capataz/s"
 )
 
 // default port for prometheus metrics server
-const metricsAddr = "0.0.0.0:8080"
-
-// newGreeter returns a worker goroutine that prints the given name every delay
-// duration of time
-func newGreeter(log *logrus.Entry, name string, delay time.Duration) c.ChildSpec {
-	ticker := time.NewTicker(delay)
-	return c.New(name, func(ctx context.Context) error {
-		for {
-			log.Infof("Hello %s", name)
-			select {
-			case <-ctx.Done():
-				return nil
-			case <-ticker.C:
-			}
-		}
-	})
-}
+const metricsHTTPAddr = "0.0.0.0:8080"
 
 // stopOnSignal will wait for the program to receive a SIGTERM and stop the
 // given supervisor
@@ -43,46 +24,72 @@ func stopOnSignal(sup s.Supervisor) {
 	fmt.Println(err)
 }
 
-func newLogEventNotifier() (*logrus.Entry, s.EventNotifier) {
-	log := logrus.New()
-	log.Out = os.Stdout
-	log.Level = logrus.DebugLevel
-	log.SetFormatter(&logrus.JSONFormatter{})
-
-	ll := log.WithFields(logrus.Fields{})
-
-	return ll, func(ev s.Event) {
-		if ev.Err() != nil {
-			ll = log.WithError(ev.Err())
-		}
-		ll.WithFields(logrus.Fields{
-			"process_runtime_name": ev.ProcessRuntimeName(),
-			"created_at":           ev.Created(),
-		}).Debug(ev.Tag().String())
-	}
-}
-
+// This program runs a prometheus metrics server and a few goroutines that print
+// greetings in the standard output. It showcases how to build by composing all
+// the components in a tree shape. All the wiring happens before we run any
+// business logic.
+//
+// This API gives restart logic for any failures that happen in the goroutines
+// and also forces us to build our application as a group of sub-systems and
+// goroutine workers.
+//
 func main() {
+	// Setup the logging mechanisms
 	log, logEventNotifier := newLogEventNotifier()
 
-	// We build a supervision tree that runs the Prometheus HTTP server
-	prometheusSpec, promEventNotifier := newPrometheusSpec("prometheus", metricsAddr)
+	// Build a supervision tree that runs the Prometheus HTTP server
+	prometheusSpec, promEventNotifier := newPrometheusSpec("prometheus", metricsHTTPAddr)
 
-	// Run the root supervisor
+	// Build a supervision tree that runs 3 greeters
+	greetersSpec := newGretterTree("greeters",
+		greeterSpec{name: "greeter1", delay: 5 * time.Second},
+		greeterSpec{name: "greeter2", delay: 7 * time.Second},
+		greeterSpec{name: "greeter3", delay: 10 * time.Second},
+	)
+
+	// We _build_ the application structure, composing all the sub-components
+	// together
 	app := s.New(
+		// The name of the topmost supervisor in our application
 		"root",
-		// Use the EventNotifier that reports metrics to Prometheus
+		// Setup the supervision system EventNotifier
+		//
+		// The EventNotifier will emit an Event everytime the supervisor:
+		//
+		// * starts a worker
+		// * stops a worker
+		// * detectes a worker failure
+		//
+		// This callback is the ideal to add traceability to our supervision system,
+		// in this particular s.EventNotifer, we log each event and send metrics to
+		// prometheus to check our application health
+		//
 		s.WithNotifier(func(ev s.Event) {
 			logEventNotifier(ev)
 			promEventNotifier(ev)
 		}),
-		// Run the prometheus supervision sub-tree
-		s.WithSubtree(prometheusSpec),
-		// Run the greeter goroutines
-		s.WithChildren(
-			newGreeter(log, "greeter1", 5*time.Second),
-			newGreeter(log, "greeter2", 7*time.Second),
-			newGreeter(log, "greeter3", 10*time.Second),
+		// When the "root" supervisor starts, it's going to start these two
+		// supervisors (sub-branches). Our root supervisor is not concerned
+		// about what these sub-systems do.
+		//
+		// In this particular example we have two sub-systems:
+		//
+		// * prometheusSpec: runs the prometheus infrastructure for our application
+		//
+		// * greetersSpec: runs a bunch of worker goroutines that greet over the
+		// terminal
+		//
+		// Given each of them is a sub-tree supervisor, if any of the leaf workers
+		// fail, they are going to get restarted by their direct supervisor. If the
+		// number of failures passes a treshold, then its going to report the error
+		// to the root supervisor, which can decide to restart the whole system, or
+		// just that particular sub-tree.
+		//
+		// In case the root supervisor error treshold is reached, then the
+		// application will fail hard.
+		s.WithSubtree(
+			prometheusSpec,
+			greetersSpec,
 		),
 	)
 
@@ -90,23 +97,40 @@ func main() {
 	//
 	// root (supervisor that restarts children gorotines)
 	// |
-	// ` prometheus (supervisor that restarts children goroutines)
+	// ` prometheus (supervisor that restarts http goroutines)
 	// | |
 	// | ` listen-and-serve (goroutine that runs http.ListenAndServe)
 	// | |
 	// | ` wait-server (goroutine that calls http.Shutdown on supervisor shutdown)
 	// |
-	// ` greeter1 (goroutine that greets on terminal)
-	// |
-	// ` greeter2 (goroutine that greets on terminal)
-	// |
-	// ` greeter3 (goroutine that greets on terminal)
+	// ` greeters (supervisor that restarts greeters goroutines)
+	//   |
+	//   ` greeter1 (goroutine that greets on terminal)
+	//   |
+	//   ` greeter2 (goroutine that greets on terminal)
+	//   |
+	//   ` greeter3 (goroutine that greets on terminal)
 	//
+	//
+	// The Start method will start each sub-component in the order above. For
+	// example, the greeters sub-tree will not start until the prometheus tree has
+	// reported it started. When using `c.New`, the start notification happens as
+	// soon as the child goroutine executes, if you use `c.NewWithStart` you can
+	// signal to the supervisor when the supervisor has started, this is useful
+	// when you need to wait for some setup before your system is "running".
 	sup, err := app.Start(context.Background())
 	if err != nil {
 		panic(err)
 	}
 
-	// We wait for signals from the OS to stop the supervisor
+	// We block, waiting for signals from the OS to stop the supervisors
+	// gracefully.
+	//
+	// When we terminate the supervisor, it will terminate each sub-tree and leaf
+	// child in the reverse order, meaning, if the child "greeter3" was the last
+	// to start, then it will be the first child that is going to be terminated.
+	//
+	// Note, goroutine termination relies on the `context.Context` given in the
+	// `c.New` call.
 	stopOnSignal(sup)
 }
